+++ src/com/chocohead/eumj/EngineMod.java
@@ -19,84 +19,38 @@ import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
+import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLConstructionEvent;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.registry.GameRegistry;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-import net.minecraftforge.oredict.RecipeSorter;
-import net.minecraftforge.oredict.RecipeSorter.Category;
 
-import buildcraft.api.BCBlocks;
-import buildcraft.api.BCItems;
-import buildcraft.api.blocks.CustomRotationHelper;
-import buildcraft.api.enums.EnumEngineType;
 import buildcraft.api.mj.MjAPI;
-import buildcraft.lib.client.guide.GuideManager;
-import buildcraft.lib.client.guide.PageLine;
-import buildcraft.lib.client.guide.loader.XmlPageLoader;
-import buildcraft.lib.client.guide.parts.GuideText;
-import buildcraft.lib.gui.GuiStack;
-import buildcraft.lib.gui.ISimpleDrawable;
-import buildcraft.transport.BCTransportItems;
 
 import ic2.api.event.TeBlockFinalCallEvent;
-import ic2.api.item.IC2Items;
 
-import ic2.core.block.BlockTileEntity;
-import ic2.core.block.TeBlockRegistry;
-import ic2.core.item.ItemIC2;
-import ic2.core.util.StackUtil;
+@Mod
+(
+	modid 		 = EngineMod.MODID,
+	name		 = EngineMod.MODNAME,
+	dependencies = EngineMod.DEPENDENCIES,
+	version		 = EngineMod.VERSION
+)
 
-import com.chocohead.eumj.item.ItemReaderMJ;
-import com.chocohead.eumj.te.Engine_TEs;
-import com.chocohead.eumj.te.TileEntityEngine;
-import com.chocohead.eumj.util.AdvEngineRecipe;
-
-@Mod(modid=MODID, name="EU-MJ Engine", dependencies="required-after:ic2;required-after:buildcraftenergy@[7.99.7];after:buildcrafttransport", version="@VERSION@")
 public final class EngineMod {
-	public static final String MODID = "eu-mj_engine";
-	public static final CreativeTabs TAB = new CreativeTabs("EU-MJ Engine") {
-		private ItemStack[] items;
-		private int ticker;
 
-		@Override
-		@SideOnly(Side.CLIENT)
-		public ItemStack getIconItemStack() {
-			if (++ticker >= 500) {
-				ticker = 0;
-			}
+	/* Meta Data */
+	public static final String MODID 		= "eu-mj_engine";
+	public static final String MODNAME 		= "EU-MJ Engine";
+	public static final String DEPENDENCIES = "required-after:ic2;required-after:buildcraftenergy@[7.99.7];after:buildcrafttransport";
+	public static final String VERSION 		= "@VERSION@";
 
-			if (items == null) {
-				items = new ItemStack[5];
+	/* Proxy */
+	@SidedProxy(clientSide = "com.chocohead.eumj.ClientProxy", serverSide = "com.chocohead.eumj.ServerProxy")
+	public static CommonProxy proxy;
 
-				for (int i = 0; i < Engine_TEs.VALUES.length; i++) {
-					items[i] = engine.getItemStack(Engine_TEs.VALUES[i]);
-				}
-			}
-
-			assert ticker / 100 < items.length;
-			return items[ticker / 100];
-		}
-
-		@Override
-		@SideOnly(Side.CLIENT)
-		public ItemStack getTabIconItem() {
-			return null; //Only normally called from getIconItemStack
-		}
-
-		@Override
-		@SideOnly(Side.CLIENT)
-		public String getTranslatedTabLabel() {
-			return MODID + ".creative_tab";
-		}
-	};
-
-	public static BlockTileEntity engine;
-	public static ItemIC2 readerMJ;
+	/** Initialization Events **/
 
 	@EventHandler
 	public void construction(FMLConstructionEvent event) {
@@ -105,7 +59,7 @@ public final class EngineMod {
 
 	@SubscribeEvent
 	public void register(TeBlockFinalCallEvent event) {
-		TeBlockRegistry.addAll(Engine_TEs.class, Engine_TEs.IDENTITY);
+		proxy.register();
 	}
 
 	@EventHandler
@@ -113,23 +67,31 @@ public final class EngineMod {
 		loadConfig(event.getSuggestedConfigurationFile());
 		event.getModLog().info("Running with "+(Conversion.MJperEU / MjAPI.MJ)+" MJ per EU or "+(MjAPI.MJ / Conversion.MJperEU)+" EU per MJ");
 
-		//Blocks
-		engine = TeBlockRegistry.get(Engine_TEs.IDENTITY);
-		engine.setCreativeTab(TAB);
-		//Items
-		if (Loader.isModLoaded("buildcrafttransport")) {
-			readerMJ = new ItemReaderMJ();
+		proxy.preInit();
+	}
+
+	@EventHandler
+	public void init(FMLInitializationEvent event) {
+		proxy.init();
+	}
+
+	@EventHandler
+	public void postInit(FMLPostInitializationEvent event) {
+		proxy.postInit();
+	}
+
+	/* Conversion Laws/Equations/Rules */
+	public static class Conversion {
+		static double MJperEU = MjAPI.MJ * 2 / 5;
+		public static double MJtoEU(long microjoules) {
+			return microjoules / MJperEU;
 		}
-
-		RecipeSorter.register(MODID+":shaped", AdvEngineRecipe.class, Category.SHAPED, "after:ic2:shaped");
-
-		if (event.getSide().isClient()) {
-			MinecraftForge.EVENT_BUS.register(MagicModelLoader.class);
-			//ModelLoader.setCustomStateMapper(engine, block -> Collections.emptyMap());
-			readerMJ.registerModels(null);
+		public static long EUtoMJ(double EU) {
+			return (long) (EU * MJperEU);
 		}
 	}
 
+	/* Import/Load and evaluate Configuration File */
 	private void loadConfig(File file) {
 		Configuration config = new Configuration(file);
 
@@ -143,113 +105,6 @@ public final class EngineMod {
 			if (config.hasChanged()) {
 				config.save();
 			}
-		}
-	}
-
-	@EventHandler
-	public void init(FMLInitializationEvent event) {
-		Engine_TEs.buildDummies(event.getSide().isClient());
-
-		if (BCBlocks.CORE_ENGINE != null) {
-			GameRegistry.addRecipe(new AdvEngineRecipe(engine.getItemStack(Engine_TEs.slow_electric_engine),
-					"B", "E", "C",
-					'B', IC2Items.getItem("re_battery"),
-					'E', new ItemStack(BCBlocks.CORE_ENGINE, 1, EnumEngineType.STONE.ordinal()),
-					'C', IC2Items.getItem("crafting", "circuit")));
-
-			GameRegistry.addRecipe(new AdvEngineRecipe(engine.getItemStack(Engine_TEs.regular_electric_engine),
-					"B", "E", "C",
-					'B', IC2Items.getItem("re_battery"),
-					'E', new ItemStack(BCBlocks.CORE_ENGINE, 1, EnumEngineType.IRON.ordinal()),
-					'C', IC2Items.getItem("crafting", "circuit")));
-
-			GameRegistry.addRecipe(new AdvEngineRecipe(engine.getItemStack(Engine_TEs.fast_electric_engine),
-					"BBB", "EPE", "CPC",
-					'B', IC2Items.getItem("advanced_re_battery"),
-					'E', new ItemStack(BCBlocks.CORE_ENGINE, 1, EnumEngineType.IRON.ordinal()),
-					'P', IC2Items.getItem("crafting", "alloy"),
-					'C', IC2Items.getItem("crafting", "circuit")));
-
-			GameRegistry.addRecipe(new AdvEngineRecipe(engine.getItemStack(Engine_TEs.quick_electric_engine),
-					"BPB", "EEE", "CPC",
-					'B', IC2Items.getItem("energy_crystal"),
-					'E', new ItemStack(BCBlocks.CORE_ENGINE, 1, EnumEngineType.IRON.ordinal()),
-					'P', IC2Items.getItem("crafting", "alloy"),
-					'C', IC2Items.getItem("crafting", "advanced_circuit")));
-
-			GameRegistry.addRecipe(new AdvEngineRecipe(engine.getItemStack(Engine_TEs.adjustable_electric_engine),
-					"BCB", "EEE", "MTM",
-					'B', IC2Items.getItem("lapotron_crystal"),
-					'E', new ItemStack(BCBlocks.CORE_ENGINE, 1, EnumEngineType.IRON.ordinal()),
-					'C', IC2Items.getItem("crafting", "advanced_circuit"),
-					'M', IC2Items.getItem("resource", "advanced_machine"),
-					'T', IC2Items.getItem("te", "hv_transformer")));
-		}
-
-		if (readerMJ != null && BCItems.CORE_GEAR_GOLD != null && BCItems.TRANSPORT_PIPE_WOOD_POWER != null) {
-			Collection<ItemStack> pipes = new HashSet<>();
-
-			for (Item pipe : new Item[] {BCTransportItems.pipePowerCobble, BCTransportItems.pipePowerStone,
-					BCTransportItems.pipePowerQuartz, BCTransportItems.pipePowerGold, BCTransportItems.pipePowerSandstone}) {
-				if (pipe != null) {
-					pipes.add(new ItemStack(pipe));
-				}
-			}
-
-			if (!pipes.isEmpty()) {
-				GameRegistry.addRecipe(new AdvEngineRecipe(new ItemStack(readerMJ),
-						" D ", "PGP", "p p",
-						'D', Items.GLOWSTONE_DUST,
-						'G', BCItems.CORE_GEAR_GOLD,
-						'P', pipes,
-						'p', BCItems.TRANSPORT_PIPE_WOOD_POWER));
-			}
-		}
-
-		//BuildCraft Lib loads pages in post-init, but it also loads first, so we do this here
-		XmlPageLoader.TAG_FACTORIES.put("engineLink", tag -> {
-			ItemStack stack = XmlPageLoader.loadItemStack(tag);
-
-			PageLine line;
-			if (StackUtil.isEmpty(stack)) {
-				line = new PageLine(1, "Missing item: "+tag, false);
-			} else {
-				ISimpleDrawable icon = new GuiStack(stack);
-				line = new PageLine(icon, icon, 1, stack.getDisplayName(), true);
-			}
-
-			return Collections.singletonList(gui -> new GuideText(gui, line) {
-				@Override
-				public PagePosition handleMouseClick(int x, int y, int width, int height, PagePosition current, int index, int mouseX, int mouseY) {
-					if (line.link && (wasHovered || wasIconHovered)) {
-						gui.openPage(GuideManager.INSTANCE.getPageFor(stack).createNew(gui));
-					}
-
-					return renderLine(current, text, x, y, width, height, -1);
-				}
-			});
-		});
-	}
-
-	@EventHandler
-	public void postInit(FMLPostInitializationEvent event) {
-		CustomRotationHelper.INSTANCE.registerHandler(engine, (world, pos, state, side) -> {
-			TileEntity te = world.getTileEntity(pos);
-
-			return te instanceof TileEntityEngine && ((TileEntityEngine) te).trySpin(side.getOpposite()) ? EnumActionResult.SUCCESS : EnumActionResult.FAIL;
-		});
-	}
-
-
-	public static class Conversion {
-		static double MJperEU = MjAPI.MJ * 2 / 5;
-
-		public static double MJtoEU(long microjoules) {
-			return microjoules / MJperEU;
-		}
-
-		public static long EUtoMJ(double EU) {
-			return (long) (EU * MJperEU);
 		}
 	}
 }
